import: refactor ingest mod
