+++ src/import/mod.rs
@@ -13,12 +13,15 @@
 //! inside TiKV because it needs to interact with raftstore.
 
 mod duplicate_detect;
+mod ingest;
 mod raft_writer;
 mod sst_service;
 
 use std::fmt::Debug;
 
 use grpcio::{RpcStatus, RpcStatusCode};
+use kvproto::errorpb;
+use sst_importer::metrics::IMPORTER_ERROR_VEC;
 pub use sst_importer::{Config, Error, Result, SstImporter, TxnSstWriter};
 
 pub use self::sst_service::ImportSstService;
@@ -48,4 +51,29 @@ macro_rules! send_rpc_response {
         };
         let _ = res.map_err(|e| warn!("send rpc response"; "err" => %e)).await;
     }};
+}
+
+// add error statistics from pb error response
+fn pb_error_inc(type_: &str, e: &errorpb::Error) {
+    let label = if e.has_not_leader() {
+        "not_leader"
+    } else if e.has_store_not_match() {
+        "store_not_match"
+    } else if e.has_region_not_found() {
+        "region_not_found"
+    } else if e.has_key_not_in_region() {
+        "key_not_in_range"
+    } else if e.has_epoch_not_match() {
+        "epoch_not_match"
+    } else if e.has_server_is_busy() {
+        "server_is_busy"
+    } else if e.has_stale_command() {
+        "stale_command"
+    } else if e.has_raft_entry_too_large() {
+        "raft_entry_too_large"
+    } else {
+        "unknown"
+    };
+
+    IMPORTER_ERROR_VEC.with_label_values(&[type_, label]).inc();
 }
