+++ src/import/sst_service.rs
@@ -1,18 +1,13 @@
 // Copyright 2018 TiKV Project Authors. Licensed under Apache-2.0.
 
 use std::{
-    collections::{HashMap, HashSet, VecDeque},
+    collections::{HashMap, VecDeque},
     convert::identity,
-    future::Future,
-    path::PathBuf,
-    sync::{
-        atomic::{AtomicU64, Ordering},
-        Arc, Mutex,
-    },
+    sync::{Arc, Mutex},
     time::{Duration, Instant as StdInstant},
 };
 
-use engine_traits::{CompactExt, MiscExt, CF_DEFAULT, CF_WRITE};
+use engine_traits::{CompactExt, CF_DEFAULT, CF_WRITE};
 use file_system::{set_io_type, IoType};
 use futures::{sink::SinkExt, stream::TryStreamExt, FutureExt, TryFutureExt};
 use grpcio::{
@@ -22,11 +17,10 @@ use kvproto::{
     encryptionpb::EncryptionMethod,
     errorpb,
     import_sstpb::{
-        Error as ImportPbError, ImportSst, Range, RawWriteRequest_oneof_chunk as RawChunk, SstMeta,
+        Error as ImportPbError, ImportSst, Range, RawWriteRequest_oneof_chunk as RawChunk,
         SuspendImportRpcRequest, SuspendImportRpcResponse, SwitchMode,
         WriteRequest_oneof_chunk as Chunk, *,
     },
-    kvrpcpb::Context,
     metapb::RegionEpoch,
 };
 use raftstore::{
@@ -37,12 +31,10 @@ use raftstore::{
 use raftstore_v2::StoreMeta;
 use resource_control::{with_resource_limiter, ResourceGroupManager};
 use sst_importer::{
-    error_inc, metrics::*, sst_importer::DownloadExt, sst_meta_to_path, Config, ConfigManager,
-    Error, Result, SstImporter,
+    error_inc, metrics::*, sst_importer::DownloadExt, Config, ConfigManager, Error, Result,
+    SstImporter,
 };
-use tikv_kv::{
-    Engine, LocalTablets, Modify, SnapContext, Snapshot, SnapshotExt, WriteData, WriteEvent,
-};
+use tikv_kv::{Engine, LocalTablets, Modify, WriteData};
 use tikv_util::{
     config::ReadableSize,
     future::{create_stream_with_buffer, paired_future_callback},
@@ -51,11 +43,11 @@ use tikv_util::{
     HandyRwLock,
 };
 use tokio::{runtime::Runtime, time::sleep};
-use txn_types::{Key, TimeStamp, WriteRef, WriteType};
+use txn_types::{Key, WriteRef, WriteType};
 
 use super::{
-    make_rpc_error,
-    raft_writer::{self, wait_write},
+    ingest::{async_snapshot, ingest, IngestLatch, SuspendDeadline},
+    make_rpc_error, pb_error_inc, raft_writer,
 };
 use crate::{
     import::duplicate_detect::DuplicateDetector,
@@ -129,7 +121,7 @@ pub struct ImportSstService<E: Engine> {
     threads: Arc<Runtime>,
     importer: Arc<SstImporter<E::Local>>,
     limiter: Limiter,
-    task_slots: Arc<Mutex<HashSet<PathBuf>>>,
+    ingest_latch: Arc<IngestLatch>,
     raft_entry_max_size: ReadableSize,
     region_info_accessor: Arc<RegionInfoAccessor>,
 
@@ -140,7 +132,7 @@ pub struct ImportSstService<E: Engine> {
     resource_manager: Option<Arc<ResourceGroupManager>>,
 
     // When less than now, don't accept any requests.
-    suspend_req_until: Arc<AtomicU64>,
+    suspend: Arc<SuspendDeadline>,
 }
 
 struct RequestCollector {
@@ -372,13 +364,13 @@ impl<E: Engine> ImportSstService<E> {
             engine,
             importer,
             limiter: Limiter::new(f64::INFINITY),
-            task_slots: Arc::new(Mutex::new(HashSet::default())),
+            ingest_latch: Arc::default(),
             raft_entry_max_size,
             region_info_accessor,
             writer,
             store_meta,
             resource_manager,
-            suspend_req_until: Arc::new(AtomicU64::new(0)),
+            suspend: Arc::default(),
         }
     }
 
@@ -392,173 +384,6 @@ impl<E: Engine> ImportSstService<E> {
 
             importer.update_config_memory_use_ratio(&cfg);
             importer.shrink_by_tick();
-        }
-    }
-
-    fn acquire_lock(task_slots: &Arc<Mutex<HashSet<PathBuf>>>, meta: &SstMeta) -> Result<bool> {
-        let mut slots = task_slots.lock().unwrap();
-        let p = sst_meta_to_path(meta)?;
-        Ok(slots.insert(p))
-    }
-
-    fn release_lock(task_slots: &Arc<Mutex<HashSet<PathBuf>>>, meta: &SstMeta) -> Result<bool> {
-        let mut slots = task_slots.lock().unwrap();
-        let p = sst_meta_to_path(meta)?;
-        Ok(slots.remove(&p))
-    }
-
-    fn async_snapshot(
-        engine: &mut E,
-        context: &Context,
-    ) -> impl Future<Output = std::result::Result<E::Snap, errorpb::Error>> {
-        let res = engine.async_snapshot(SnapContext {
-            pb_ctx: context,
-            ..Default::default()
-        });
-        async move {
-            res.await.map_err(|e| {
-                let err: storage::Error = e.into();
-                if let Some(e) = extract_region_error_from_error(&err) {
-                    e
-                } else {
-                    let mut e = errorpb::Error::default();
-                    e.set_message(format!("{}", err));
-                    e
-                }
-            })
-        }
-    }
-
-    fn check_write_stall(&self, region_id: u64) -> Option<errorpb::Error> {
-        let tablet = match self.tablets.get(region_id) {
-            Some(tablet) => tablet,
-            None => {
-                let mut errorpb = errorpb::Error::default();
-                errorpb.set_message(format!("region {} not found", region_id));
-                errorpb.mut_region_not_found().set_region_id(region_id);
-                return Some(errorpb);
-            }
-        };
-
-        let reject_error = |region_id: Option<u64>| -> Option<errorpb::Error> {
-            let mut errorpb = errorpb::Error::default();
-            let err = if let Some(id) = region_id {
-                format!("too many sst files are ingesting for region {}", id)
-            } else {
-                "too many sst files are ingesting".to_string()
-            };
-            let mut server_is_busy_err = errorpb::ServerIsBusy::default();
-            server_is_busy_err.set_reason(err.clone());
-            errorpb.set_message(err);
-            errorpb.set_server_is_busy(server_is_busy_err);
-            Some(errorpb)
-        };
-
-        // store_meta being Some means it is v2
-        if let Some(ref store_meta) = self.store_meta {
-            if let Some((region, _)) = store_meta.lock().unwrap().regions.get(&region_id) {
-                if !self.importer.region_in_import_mode(region)
-                    && tablet.ingest_maybe_slowdown_writes(CF_WRITE).expect("cf")
-                {
-                    return reject_error(Some(region_id));
-                }
-            } else {
-                let mut errorpb = errorpb::Error::default();
-                errorpb.set_message(format!("region {} not found", region_id));
-                errorpb.mut_region_not_found().set_region_id(region_id);
-                return Some(errorpb);
-            }
-        } else if self.importer.get_mode() == SwitchMode::Normal
-            && tablet.ingest_maybe_slowdown_writes(CF_WRITE).expect("cf")
-        {
-            match tablet.get_sst_key_ranges(CF_WRITE, 0) {
-                Ok(l0_sst_ranges) => {
-                    warn!(
-                        "sst ingest is too slow";
-                        "sst_ranges" => ?l0_sst_ranges,
-                    );
-                }
-                Err(e) => {
-                    error!("get sst key ranges failed"; "err" => ?e);
-                }
-            }
-            return reject_error(None);
-        }
-
-        None
-    }
-
-    fn ingest_files(
-        &mut self,
-        mut context: Context,
-        label: &'static str,
-        ssts: Vec<SstMeta>,
-    ) -> impl Future<Output = Result<IngestResponse>> {
-        let snapshot_res = Self::async_snapshot(&mut self.engine, &context);
-        let engine = self.engine.clone();
-        let importer = self.importer.clone();
-        async move {
-            // check api version
-            if !importer.as_ref().check_api_version(&ssts)? {
-                return Err(Error::IncompatibleApiVersion);
-            }
-
-            let mut resp = IngestResponse::default();
-            let res = match snapshot_res.await {
-                Ok(snap) => snap,
-                Err(e) => {
-                    pb_error_inc(label, &e);
-                    resp.set_error(e);
-                    return Ok(resp);
-                }
-            };
-
-            fail_point!("import::sst_service::ingest");
-            // Here we shall check whether the file has been ingested before. This operation
-            // must execute after geting a snapshot from raftstore to make sure that the
-            // current leader has applied to current term.
-            for sst in ssts.iter() {
-                if !importer.exist(sst) {
-                    warn!(
-                        "sst [{:?}] not exist. we may retry an operation that has already succeeded",
-                        sst
-                    );
-                    let mut errorpb = errorpb::Error::default();
-                    let err = "The file which would be ingested doest not exist.";
-                    let stale_err = errorpb::StaleCommand::default();
-                    errorpb.set_message(err.to_string());
-                    errorpb.set_stale_command(stale_err);
-                    resp.set_error(errorpb);
-                    return Ok(resp);
-                }
-            }
-            let modifies = ssts
-                .iter()
-                .map(|s| Modify::Ingest(Box::new(s.clone())))
-                .collect();
-            context.set_term(res.ext().get_term().unwrap().into());
-            let region_id = context.get_region_id();
-            let res = engine.async_write(
-                &context,
-                WriteData::from_modifies(modifies),
-                WriteEvent::BASIC_EVENT,
-                None,
-            );
-
-            let mut resp = IngestResponse::default();
-            if let Err(e) = wait_write(res).await {
-                if let Some(e) = extract_region_error_from_error(&e) {
-                    pb_error_inc(label, &e);
-                    resp.set_error(e);
-                } else {
-                    IMPORTER_ERROR_VEC
-                        .with_label_values(&[label, "unknown"])
-                        .inc();
-                    resp.mut_error()
-                        .set_message(format!("[region {}] ingest failed: {:?}", region_id, e));
-                }
-            }
-            Ok(resp)
         }
     }
 
@@ -641,47 +466,6 @@ impl<E: Engine> ImportSstService<E> {
         futures::future::try_join_all(inflight_futures).await?;
 
         Ok(range)
-    }
-
-    /// Check whether we should suspend the current request.
-    fn check_suspend(&self) -> Result<()> {
-        let now = TimeStamp::physical_now();
-        let suspend_until = self.suspend_req_until.load(Ordering::SeqCst);
-        if now < suspend_until {
-            Err(Error::Suspended {
-                time_to_lease_expire: Duration::from_millis(suspend_until - now),
-            })
-        } else {
-            Ok(())
-        }
-    }
-
-    /// suspend requests for a period.
-    ///
-    /// # returns
-    ///
-    /// whether for now, the requests has already been suspended.
-    pub fn suspend_requests(&self, for_time: Duration) -> bool {
-        let now = TimeStamp::physical_now();
-        let last_suspend_until = self.suspend_req_until.load(Ordering::SeqCst);
-        let suspended = now < last_suspend_until;
-        let suspend_until = TimeStamp::physical_now() + for_time.as_millis() as u64;
-        self.suspend_req_until
-            .store(suspend_until, Ordering::SeqCst);
-        suspended
-    }
-
-    /// allow all requests to enter.
-    ///
-    /// # returns
-    ///
-    /// whether requests has already been previously suspended.
-    pub fn allow_requests(&self) -> bool {
-        let now = TimeStamp::physical_now();
-        let last_suspend_until = self.suspend_req_until.load(Ordering::SeqCst);
-        let suspended = now < last_suspend_until;
-        self.suspend_req_until.store(0, Ordering::SeqCst);
-        suspended
     }
 }
 
@@ -955,6 +739,7 @@ impl<E: Engine> ImportSst for ImportSstService<E> {
         let acquire_lease_task = async move {
             let mut resp = LeaseResponse::default();
             let now = StdInstant::now();
+            info!("dbg rpc lease"; "req" => ?req);
 
             // Acquire leases.
             for acquire_lease in req.get_acquire() {
@@ -962,6 +747,7 @@ impl<E: Engine> ImportSst for ImportSstService<E> {
                 let uuid = acquire_lease.get_lease().get_uuid();
                 let lease_secs =
                     std::cmp::min(acquire_lease.get_ttl(), SST_LEASE_DURATION_MAX_SECS);
+                info!("dbg rpc lease"; "region_id" => region_id, "lease_secs" => lease_secs);
                 let deadline = now + Duration::from_secs(lease_secs);
                 match import.acquire_lease(region_id, uuid, deadline) {
                     Ok(_) => {
@@ -1218,47 +1004,34 @@ impl<E: Engine> ImportSst for ImportSstService<E> {
     /// CleanupSstWorker.
     fn ingest(
         &mut self,
-        ctx: RpcContext<'_>,
+        _: RpcContext<'_>,
         mut req: IngestRequest,
         sink: UnarySink<IngestResponse>,
     ) {
         let label = "ingest";
         let timer = Instant::now_coarse();
-        let mut resp = IngestResponse::default();
+        let import = self.importer.clone();
+        let engine = self.engine.clone();
+        let suspend = self.suspend.clone();
+        let tablets = self.tablets.clone();
+        let store_meta = self.store_meta.clone();
+        let ingest_latch = self.ingest_latch.clone();
 
-        if let Err(err) = self.check_suspend() {
-            resp.set_error(ImportPbError::from(err).take_store_error());
-            ctx.spawn(async move { crate::send_rpc_response!(Ok(resp), sink, label, timer) });
-            return;
-        }
-
-        let region_id = req.get_context().get_region_id();
-        if let Some(errorpb) = self.check_write_stall(region_id) {
-            resp.set_error(errorpb);
-            ctx.spawn(
-                sink.success(resp)
-                    .unwrap_or_else(|e| warn!("send rpc failed"; "err" => %e)),
-            );
-            return;
-        }
-
-        let mut errorpb = errorpb::Error::default();
-        if !Self::acquire_lock(&self.task_slots, req.get_sst()).unwrap_or(false) {
-            errorpb.set_message(Error::FileConflict.to_string());
-            resp.set_error(errorpb);
-            ctx.spawn(
-                sink.success(resp)
-                    .unwrap_or_else(|e| warn!("send rpc failed"; "err" => %e)),
-            );
-            return;
-        }
-
-        let task_slots = self.task_slots.clone();
-        let meta = req.take_sst();
-        let f = self.ingest_files(req.take_context(), label, vec![meta.clone()]);
         let handle_task = async move {
-            let res = f.await;
-            Self::release_lock(&task_slots, &meta).unwrap();
+            let mut multi_ingest = MultiIngestRequest::default();
+            multi_ingest.set_context(req.take_context());
+            multi_ingest.mut_ssts().push(req.take_sst());
+            let res = ingest(
+                multi_ingest,
+                engine,
+                &suspend,
+                &tablets,
+                &store_meta,
+                &import,
+                &ingest_latch,
+                label,
+            )
+            .await;
             crate::send_rpc_response!(res, sink, label, timer);
         };
         self.threads.spawn(handle_task);
@@ -1267,54 +1040,31 @@ impl<E: Engine> ImportSst for ImportSstService<E> {
     /// Ingest multiple files by sending a raft command to raftstore.
     fn multi_ingest(
         &mut self,
-        ctx: RpcContext<'_>,
-        mut req: MultiIngestRequest,
+        _: RpcContext<'_>,
+        req: MultiIngestRequest,
         sink: UnarySink<IngestResponse>,
     ) {
         let label = "multi-ingest";
         let timer = Instant::now_coarse();
-        let mut resp = IngestResponse::default();
-        if let Err(err) = self.check_suspend() {
-            resp.set_error(ImportPbError::from(err).take_store_error());
-            ctx.spawn(async move { crate::send_rpc_response!(Ok(resp), sink, label, timer) });
-            return;
-        }
+        let import = self.importer.clone();
+        let engine = self.engine.clone();
+        let suspend = self.suspend.clone();
+        let tablets = self.tablets.clone();
+        let store_meta = self.store_meta.clone();
+        let ingest_latch = self.ingest_latch.clone();
 
-        if let Some(errorpb) = self.check_write_stall(req.get_context().get_region_id()) {
-            resp.set_error(errorpb);
-            ctx.spawn(
-                sink.success(resp)
-                    .unwrap_or_else(|e| warn!("send rpc failed"; "err" => %e)),
-            );
-            return;
-        }
-
-        let mut errorpb = errorpb::Error::default();
-        let mut metas = vec![];
-        for sst in req.get_ssts() {
-            if Self::acquire_lock(&self.task_slots, sst).unwrap_or(false) {
-                metas.push(sst.clone());
-            }
-        }
-        if metas.len() < req.get_ssts().len() {
-            for m in metas {
-                Self::release_lock(&self.task_slots, &m).unwrap();
-            }
-            errorpb.set_message(Error::FileConflict.to_string());
-            resp.set_error(errorpb);
-            ctx.spawn(
-                sink.success(resp)
-                    .unwrap_or_else(|e| warn!("send rpc failed"; "err" => %e)),
-            );
-            return;
-        }
-        let task_slots = self.task_slots.clone();
-        let f = self.ingest_files(req.take_context(), label, req.take_ssts().into());
         let handle_task = async move {
-            let res = f.await;
-            for m in metas {
-                Self::release_lock(&task_slots, &m).unwrap();
-            }
+            let res = ingest(
+                req,
+                engine,
+                &suspend,
+                &tablets,
+                &store_meta,
+                &import,
+                &ingest_latch,
+                label,
+            )
+            .await;
             crate::send_rpc_response!(res, sink, label, timer);
         };
         self.threads.spawn(handle_task);
@@ -1424,7 +1174,7 @@ impl<E: Engine> ImportSst for ImportSstService<E> {
             Some(request.take_end_key())
         };
         let key_only = request.get_key_only();
-        let snap_res = Self::async_snapshot(&mut self.engine, &context);
+        let snap_res = async_snapshot(&mut self.engine, &context);
         let handle_task = async move {
             let res = snap_res.await;
             let snapshot = match res {
@@ -1504,40 +1254,16 @@ impl<E: Engine> ImportSst for ImportSstService<E> {
 
         let suspended = if req.should_suspend_imports {
             info!("suspend incoming import RPCs."; "for_second" => req.get_duration_in_secs(), "caller" => req.get_caller());
-            self.suspend_requests(Duration::from_secs(req.get_duration_in_secs()))
+            self.suspend
+                .suspend_requests(Duration::from_secs(req.get_duration_in_secs()))
         } else {
             info!("allow incoming import RPCs."; "caller" => req.get_caller());
-            self.allow_requests()
+            self.suspend.allow_requests()
         };
         let mut resp = SuspendImportRpcResponse::default();
         resp.set_already_suspended(suspended);
         ctx.spawn(async move { send_rpc_response!(Ok(resp), sink, label, timer) });
     }
-}
-
-// add error statistics from pb error response
-fn pb_error_inc(type_: &str, e: &errorpb::Error) {
-    let label = if e.has_not_leader() {
-        "not_leader"
-    } else if e.has_store_not_match() {
-        "store_not_match"
-    } else if e.has_region_not_found() {
-        "region_not_found"
-    } else if e.has_key_not_in_region() {
-        "key_not_in_range"
-    } else if e.has_epoch_not_match() {
-        "epoch_not_match"
-    } else if e.has_server_is_busy() {
-        "server_is_busy"
-    } else if e.has_stale_command() {
-        "stale_command"
-    } else if e.has_raft_entry_too_large() {
-        "raft_entry_too_large"
-    } else {
-        "unknown"
-    };
-
-    IMPORTER_ERROR_VEC.with_label_values(&[type_, label]).inc();
 }
 
 fn write_needs_restore(write: &[u8]) -> bool {
